+++ b2/templates.go
@@ -326,6 +326,12 @@ func getLifecycleRulesElem(ds bool) *schema.Resource {
 				Computed:    If(ds, true, false),
 				Optional:    If(ds, false, true),
 			},
+			"days_from_starting_to_canceling_unfinished_large_files": {
+				Description: "It cancels any unfinished large file versions after a given number of days.",
+				Type:        schema.TypeInt,
+				Computed:    If(ds, true, false),
+				Optional:    If(ds, false, true),
+			},
 		},
 	}
 }